gemspec: Move rake to development dependencies
`rake` is not really a runtime dependency. The way it was previously phrased meant that all projects using this gem would include Rake as a _runtime_ dependency also, which increased their footprint unecessarily.

This PR moves it to development dependencies, which is where such gems are normally located. `bundle install` in this repo will still give you `rake`, but `bundle install` in dependers will not necessarily do so.