Allow `http_ssl_version: nil` to use system default
Otherwise we have to keep on top of SSL versions, and might get behind (stagnate) and end up using insecure communications. The option to use the system default is not possible when `nil` options are filtered out, so instead of removing that filter, this commit detects a nil option and overrides the default after the merge and instance_variable_set loop.