Remove json gem dependency
json is bundled as part of the Ruby stdlib. While active_support used to
include json as part of their gemspec, this is no longer the case.

If for some reason a consumer of the postmark API wants to include the
json gem in their bundle, they can do so, but consumers shouldn't be
forced to include it.