Minimal support for issuing warnings for deprecated classes  (#137)
## What

Outputs a warning message when using one of the deprecated error classes
is referenced:

```
DEPRECATION WARNING: Postmark::DeliveryError is deprecated.
Use Postmark::Error instead. 
```

Replacement for https://github.com/ActiveCampaign/postmark-gem/pull/136,
this homegrown version only supports deprecating constants, but is free
of the `active_support` dependency.

## How

We're providing basic support for silencing warnings, as having noise in
the tests quickly can turn into a broken window situation, and the same
goes for any project that might be depending on this gem.